Refactor dashboard page into modular components
- Decomposed the main dashboard page into smaller, reusable components:
  - StatCards, PerformanceSection, PerformanceSummary, CourseOverview, AnnouncementsCard, BottomSection, CanvasSection, DashboardHeader
- Moved data fetching and aggregation logic to a custom hook (useDashboardData)
- Updated imports to use a central index for dashboard components
- Improved type safety and maintainability
- Cleaned up and modernized the dashboard code structure
