+++ components/dashboard/canvas-integration.tsx
