+++ components/dashboard/index.ts
@@ -1,0 +1,12 @@
+export * from './AnnouncementsCard';
+export * from './BottomSection';
+export * from './CanvasSection';
+export * from './CourseOverview';
+export * from './DashboardHeader';
+export * from './PerformanceSection';
+export * from './PerformanceSummary';
+export * from './StatCards'; 
+export * from './student-leaderboard';
+export * from './canvas-integration';
+export * from './performance-graph';
+
