+++ components/dashboard/student-details.tsx
